add glider

+++ ueaglider/data/db_classes.py
@@ -57,7 +57,7 @@ class Dives(SqlAlchemyBase):
     Longitude = sqlalchemy.Column(sqlalchemy.FLOAT)
     Status: str = sqlalchemy.Column(sqlalchemy.TEXT, nullable=True)
     ReceivedDate = sqlalchemy.Column(sqlalchemy.DATETIME, nullable=True)
-    Elevation = sqlalchemy.Column(sqlalchemy.FLOAT, nullable=True)
+    Elevation = sqlalchemy.Column(sqlalchemy.INT, nullable=True)
     # Mission relationship
     MissionID: str = sqlalchemy.Column(sqlalchemy.String, sqlalchemy.ForeignKey("Missions.Number"))
     mission = orm.relation('Missions')
